requirements.txt
ipykernel
pandas
numpy
scikit-learn
duckduckgo-search
git+https://github.com/openai/swarm.git
python-dotenv
faiss-gpu